count number of internal vs external (to msft) users
