+++ src/app/shared/appTelemetry/appInsightsClient.ts
@@ -1,6 +1,7 @@
 import { ApplicationInsights } from '@microsoft/applicationinsights-web';
 import { TELEMETRY_EVENTS } from '../../constants/telemetry';
 import { appConfig } from '../../../appConfig/appConfig';
+import { isMicrosoftInternalDomain } from '../utils/utils';
 
 export class AppInsightsClient {
     private static instance: ApplicationInsights;
@@ -13,9 +14,7 @@ export class AppInsightsClient {
             } });
             try {
                 appInsights.loadAppInsights();
-                // tslint:disable-next-line:no-console
-                console.log('INTERNAL: ' + verifyMicrosoftInternalDomain());
-                appInsights.trackEvent({name: TELEMETRY_EVENTS.INIT});
+                appInsights.trackEvent({name: TELEMETRY_EVENTS.INTERNAL_USER}, {isInternal: isMicrosoftInternalDomain()});
                 AppInsightsClient.instance = appInsights;
             } catch (e) {
                 // tslint:disable-next-line:no-console
@@ -26,15 +25,3 @@ export class AppInsightsClient {
         return AppInsightsClient.instance;
     }
 }
-
-// TODO: move to a better place
-const verifyMicrosoftInternalDomain = (): boolean => {
-    const msftInternalDomains = ['redmond.corp.microsoft.com', 'northamerica.corp.microsoft.com', 'fareast.corp.microsoft.com', 'ntdev.corp.microsoft.com', 'wingroup.corp.microsoft.com', 'southpacific.corp.microsoft.com', 'wingroup.windeploy.ntdev.microsoft.com', 'ddnet.microsoft.com', 'europe.corp.microsoft.com'];
-    const userDnsDomain = process.env.userdnsdomain;
-    if (!userDnsDomain) {
-        return false;
-    }
-
-    const domain = userDnsDomain.toLowerCase();
-    return msftInternalDomains.some(msftDomain => domain === msftDomain);
-};
